sample: reorganize ui package into sub-packages with internal visibility
- ui/core/ — Theme, Paddings, AppHeader (shared primitives)
- ui/sensors/ — SensorList, SensorEntryProvider, SenseyRadioButton,
  SenseyRadioButtonWithTouchArea, ResultArea, ScrollbarThumb
- ui/nav/ — MainScreen, GestureNavBar, GestureNavRail,
  GestureNavRailLayout, GestureCompactLayout
- ui/tabs/ — GestureGroupTab (unchanged)
- Mark all composables internal except SensorItem, sensorEntryProvider,
  MainScreen, and core types needed cross-package
