+++ sample/src/main/java/com/github/nisrulz/senseysample/ui/core/Paddings.kt
@@ -1,4 +1,4 @@
-package com.github.nisrulz.senseysample.ui
+package com.github.nisrulz.senseysample.ui.core
 
 import androidx.compose.ui.unit.dp
