+++ sample/src/main/java/com/github/nisrulz/senseysample/ui/sensors/SensorEntryProvider.kt
@@ -1,4 +1,4 @@
-package com.github.nisrulz.senseysample.ui
+package com.github.nisrulz.senseysample.ui.sensors
 
 import androidx.compose.runtime.Composable
 import androidx.navigation3.runtime.NavEntry
@@ -6,7 +6,7 @@ import com.github.nisrulz.senseysample.navigation.GestureGroup
 import com.github.nisrulz.senseysample.ui.tabs.GestureGroupTab
 
 @Composable
-internal fun sensorEntryProvider(
+fun sensorEntryProvider(
     selectedSensor: String?,
     sensorResults: Map<String, String>,
     onSensorSelected: (String) -> Unit,
